requirements.txt
python-whois
dnspython
requests
python-nmap
jinja2
